@@ -16,13 +16,12 @@ package collector
 import (
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
+	"io"
+	"log/slog"
 	"net/http"
 	"net/url"
 	"path"
 
-	"github.com/go-kit/log"
-	"github.com/go-kit/log/level"
 	"github.com/prometheus/client_golang/prometheus"
 )
 
@@ -42,35 +41,20 @@ var (
 
 // DataStream Information Struct
 type DataStream struct {
-	logger log.Logger
+	logger *slog.Logger
 	client *http.Client
 	url    *url.URL
-
-	up                              prometheus.Gauge
-	totalScrapes, jsonParseFailures prometheus.Counter
 
 	dataStreamMetrics []*dataStreamMetric
 }
 
 // NewDataStream defines DataStream Prometheus metrics
-func NewDataStream(logger log.Logger, client *http.Client, url *url.URL) *DataStream {
+func NewDataStream(logger *slog.Logger, client *http.Client, url *url.URL) *DataStream {
 	return &DataStream{
 		logger: logger,
 		client: client,
 		url:    url,
 
-		up: prometheus.NewGauge(prometheus.GaugeOpts{
-			Name: prometheus.BuildFQName(namespace, "data_stream_stats", "up"),
-			Help: "Was the last scrape of the ElasticSearch Data Stream stats endpoint successful.",
-		}),
-		totalScrapes: prometheus.NewCounter(prometheus.CounterOpts{
-			Name: prometheus.BuildFQName(namespace, "data_stream_stats", "total_scrapes"),
-			Help: "Current total ElasticSearch Data STream scrapes.",
-		}),
-		jsonParseFailures: prometheus.NewCounter(prometheus.CounterOpts{
-			Name: prometheus.BuildFQName(namespace, "data_stream_stats", "json_parse_failures"),
-			Help: "Number of errors while parsing JSON.",
-		}),
 		dataStreamMetrics: []*dataStreamMetric{
 			{
 				Type: prometheus.CounterValue,
@@ -105,10 +89,6 @@ func (ds *DataStream) Describe(ch chan<- *prometheus.Desc) {
 	for _, metric := range ds.dataStreamMetrics {
 		ch <- metric.Desc
 	}
-
-	ch <- ds.up.Desc()
-	ch <- ds.totalScrapes.Desc()
-	ch <- ds.jsonParseFailures.Desc()
 }
 
 func (ds *DataStream) fetchAndDecodeDataStreamStats() (DataStreamStatsResponse, error) {
@@ -125,8 +105,8 @@ func (ds *DataStream) fetchAndDecodeDataStreamStats() (DataStreamStatsResponse, 
 	defer func() {
 		err = res.Body.Close()
 		if err != nil {
-			_ = level.Warn(ds.logger).Log(
-				"msg", "failed to close http.Client",
+			ds.logger.Warn(
+				"failed to close http.Client",
 				"err", err,
 			)
 		}
@@ -136,14 +116,12 @@ func (ds *DataStream) fetchAndDecodeDataStreamStats() (DataStreamStatsResponse, 
 		return dsr, fmt.Errorf("HTTP Request failed with code %d", res.StatusCode)
 	}
 
-	bts, err := ioutil.ReadAll(res.Body)
+	bts, err := io.ReadAll(res.Body)
 	if err != nil {
-		ds.jsonParseFailures.Inc()
 		return dsr, err
 	}
 
 	if err := json.Unmarshal(bts, &dsr); err != nil {
-		ds.jsonParseFailures.Inc()
 		return dsr, err
 	}
 
@@ -152,24 +130,15 @@ func (ds *DataStream) fetchAndDecodeDataStreamStats() (DataStreamStatsResponse, 
 
 // Collect gets DataStream metric values
 func (ds *DataStream) Collect(ch chan<- prometheus.Metric) {
-	ds.totalScrapes.Inc()
-	defer func() {
-		ch <- ds.up
-		ch <- ds.totalScrapes
-		ch <- ds.jsonParseFailures
-	}()
 
 	dataStreamStatsResp, err := ds.fetchAndDecodeDataStreamStats()
 	if err != nil {
-		ds.up.Set(0)
-		_ = level.Warn(ds.logger).Log(
-			"msg", "failed to fetch and decode data stream stats",
+		ds.logger.Warn(
+			"failed to fetch and decode data stream stats",
 			"err", err,
 		)
 		return
 	}
-
-	ds.up.Set(1)
 
 	for _, metric := range ds.dataStreamMetrics {
 		for _, dataStream := range dataStreamStatsResp.DataStreamStats {
