@@ -16,13 +16,12 @@ package collector
 import (
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
+	"io"
+	"log/slog"
 	"net/http"
 	"net/url"
 	"path"
 
-	"github.com/go-kit/log"
-	"github.com/go-kit/log/level"
 	"github.com/prometheus/client_golang/prometheus"
 )
 
@@ -35,13 +34,9 @@ type ilmMetric struct {
 
 // Index Lifecycle Management information object
 type IlmIndiciesCollector struct {
-	logger log.Logger
+	logger *slog.Logger
 	client *http.Client
 	url    *url.URL
-
-	up                prometheus.Gauge
-	totalScrapes      prometheus.Counter
-	jsonParseFailures prometheus.Counter
 
 	ilmMetric ilmMetric
 }
@@ -64,7 +59,7 @@ var (
 )
 
 // NewIlmIndicies defines Index Lifecycle Management Prometheus metrics
-func NewIlmIndicies(logger log.Logger, client *http.Client, url *url.URL) *IlmIndiciesCollector {
+func NewIlmIndicies(logger *slog.Logger, client *http.Client, url *url.URL) *IlmIndiciesCollector {
 	subsystem := "ilm_index"
 
 	return &IlmIndiciesCollector{
@@ -72,18 +67,6 @@ func NewIlmIndicies(logger log.Logger, client *http.Client, url *url.URL) *IlmIn
 		client: client,
 		url:    url,
 
-		up: prometheus.NewGauge(prometheus.GaugeOpts{
-			Name: prometheus.BuildFQName(namespace, subsystem, "up"),
-			Help: "Was the last scrape of the ElasticSearch ILM endpoint successful.",
-		}),
-		totalScrapes: prometheus.NewCounter(prometheus.CounterOpts{
-			Name: prometheus.BuildFQName(namespace, subsystem, "total_scrapes"),
-			Help: "Current total ElasticSearch ILM scrapes.",
-		}),
-		jsonParseFailures: prometheus.NewCounter(prometheus.CounterOpts{
-			Name: prometheus.BuildFQName(namespace, subsystem, "json_parse_failures"),
-			Help: "Number of errors while parsing JSON.",
-		}),
 		ilmMetric: ilmMetric{
 			Type: prometheus.GaugeValue,
 			Desc: prometheus.NewDesc(
@@ -100,9 +83,6 @@ func NewIlmIndicies(logger log.Logger, client *http.Client, url *url.URL) *IlmIn
 // Describe adds metrics description
 func (i *IlmIndiciesCollector) Describe(ch chan<- *prometheus.Desc) {
 	ch <- i.ilmMetric.Desc
-	ch <- i.up.Desc()
-	ch <- i.totalScrapes.Desc()
-	ch <- i.jsonParseFailures.Desc()
 }
 
 func (i *IlmIndiciesCollector) fetchAndDecodeIlm() (IlmResponse, error) {
@@ -120,8 +100,8 @@ func (i *IlmIndiciesCollector) fetchAndDecodeIlm() (IlmResponse, error) {
 	defer func() {
 		err = res.Body.Close()
 		if err != nil {
-			_ = level.Warn(i.logger).Log(
-				"msg", "failed to close http.Client",
+			i.logger.Warn(
+				"failed to close http.Client",
 				"err", err,
 			)
 		}
@@ -131,14 +111,12 @@ func (i *IlmIndiciesCollector) fetchAndDecodeIlm() (IlmResponse, error) {
 		return ir, fmt.Errorf("HTTP Request failed with code %d", res.StatusCode)
 	}
 
-	bts, err := ioutil.ReadAll(res.Body)
+	bts, err := io.ReadAll(res.Body)
 	if err != nil {
-		i.jsonParseFailures.Inc()
 		return ir, err
 	}
 
 	if err := json.Unmarshal(bts, &ir); err != nil {
-		i.jsonParseFailures.Inc()
 		return ir, err
 	}
 
@@ -154,24 +132,15 @@ func bool2int(managed bool) float64 {
 
 // Collect pulls metric values from Elasticsearch
 func (i *IlmIndiciesCollector) Collect(ch chan<- prometheus.Metric) {
-	defer func() {
-		ch <- i.up
-		ch <- i.totalScrapes
-		ch <- i.jsonParseFailures
-	}()
-
 	// indices
 	ilmResp, err := i.fetchAndDecodeIlm()
 	if err != nil {
-		i.up.Set(0)
-		_ = level.Warn(i.logger).Log(
-			"msg", "failed to fetch and decode ILM stats",
+		i.logger.Warn(
+			"failed to fetch and decode ILM stats",
 			"err", err,
 		)
 		return
 	}
-	i.totalScrapes.Inc()
-	i.up.Set(1)
 
 	for indexName, indexIlm := range ilmResp.Indices {
 		ch <- prometheus.MustNewConstMetric(
