@@ -16,20 +16,19 @@ package collector
 import (
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
+	"io"
+	"log/slog"
 	"net/http"
 	"net/url"
 	"path"
 	"strconv"
 
-	"github.com/go-kit/log"
-	"github.com/go-kit/log/level"
 	"github.com/prometheus/client_golang/prometheus"
 )
 
 // IndicesSettings information struct
 type IndicesSettings struct {
-	logger log.Logger
+	logger *slog.Logger
 	client *http.Client
 	url    *url.URL
 
@@ -43,6 +42,7 @@ type IndicesSettings struct {
 var (
 	defaultIndicesTotalFieldsLabels = []string{"index"}
 	defaultTotalFieldsValue         = 1000 //es default configuration for total fields
+	defaultDateCreation             = 0    //es index default creation date
 )
 
 type indicesSettingsMetric struct {
@@ -52,7 +52,7 @@ type indicesSettingsMetric struct {
 }
 
 // NewIndicesSettings defines Indices Settings Prometheus metrics
-func NewIndicesSettings(logger log.Logger, client *http.Client, url *url.URL) *IndicesSettings {
+func NewIndicesSettings(logger *slog.Logger, client *http.Client, url *url.URL) *IndicesSettings {
 	return &IndicesSettings{
 		logger: logger,
 		client: client,
@@ -105,6 +105,21 @@ func NewIndicesSettings(logger log.Logger, client *http.Client, url *url.URL) *I
 					return val
 				},
 			},
+			{
+				Type: prometheus.GaugeValue,
+				Desc: prometheus.NewDesc(
+					prometheus.BuildFQName(namespace, "indices_settings", "creation_timestamp_seconds"),
+					"index setting creation_date",
+					defaultIndicesTotalFieldsLabels, nil,
+				),
+				Value: func(indexSettings Settings) float64 {
+					val, err := strconv.ParseFloat(indexSettings.IndexInfo.CreationDate, 64)
+					if err != nil {
+						return float64(defaultDateCreation)
+					}
+					return val / 1000.0
+				},
+			},
 		},
 	}
 }
@@ -127,8 +142,8 @@ func (cs *IndicesSettings) getAndParseURL(u *url.URL, data interface{}) error {
 	defer func() {
 		err = res.Body.Close()
 		if err != nil {
-			_ = level.Warn(cs.logger).Log(
-				"msg", "failed to close http.Client",
+			cs.logger.Warn(
+				"failed to close http.Client",
 				"err", err,
 			)
 		}
@@ -138,7 +153,7 @@ func (cs *IndicesSettings) getAndParseURL(u *url.URL, data interface{}) error {
 		return fmt.Errorf("HTTP Request failed with code %d", res.StatusCode)
 	}
 
-	bts, err := ioutil.ReadAll(res.Body)
+	bts, err := io.ReadAll(res.Body)
 	if err != nil {
 		cs.jsonParseFailures.Inc()
 		return err
@@ -179,8 +194,8 @@ func (cs *IndicesSettings) Collect(ch chan<- prometheus.Metric) {
 	if err != nil {
 		cs.readOnlyIndices.Set(0)
 		cs.up.Set(0)
-		_ = level.Warn(cs.logger).Log(
-			"msg", "failed to fetch and decode cluster settings stats",
+		cs.logger.Warn(
+			"failed to fetch and decode cluster settings stats",
 			"err", err,
 		)
 		return
