@@ -14,13 +14,16 @@
 package collector
 
 import (
-	"fmt"
+	"io"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
+	"os"
+	"strings"
 	"testing"
 
-	"github.com/go-kit/log"
+	"github.com/prometheus/client_golang/prometheus/testutil"
+	"github.com/prometheus/common/promslog"
 )
 
 func TestMapping(t *testing.T) {
@@ -66,315 +69,58 @@ func TestMapping(t *testing.T) {
 	    }
 	}'*/
 	//  curl http://localhost:9200/_all/_mapping
-	tcs := map[string]string{
-		"7.8.0": `{
-			"facebook": {
-			  "mappings": {
-				"properties": {
-				  "contact": {
-					"properties": {
-					  "email": {
-						"type": "text",
-						"fields": {
-						  "raw": {
-							"type": "keyword"
-						  }
-						}
-					  },
-					  "phone": {
-						"type": "text"
-					  }
-					}
-				  },
-				  "name": {
-					"type": "text",
-					"fields": {
-					  "raw": {
-						"type": "keyword"
-					  }
-					}
-				  }
-				}
-			  }
-			},
-			"twitter": {
-			  "mappings": {
-				"properties": {
-				  "email": {
-					"type": "keyword"
-				  },
-				  "phone": {
-					"type": "keyword"
-				  }
-				}
-			  }
-			}
-		  }`,
+	tests := []struct {
+		name string
+		file string
+		want string
+	}{
+		{
+			name: "7.8.0",
+			file: "../fixtures/indices_mappings/7.8.0.json",
+			want: `
+# HELP elasticsearch_indices_mappings_stats_fields Current number fields within cluster.
+# TYPE elasticsearch_indices_mappings_stats_fields gauge
+elasticsearch_indices_mappings_stats_fields{index="facebook"} 6
+elasticsearch_indices_mappings_stats_fields{index="twitter"} 2
+			`,
+		},
+		{
+			name: "counts",
+			file: "../fixtures/indices_mappings/counts.json",
+			want: `
+# HELP elasticsearch_indices_mappings_stats_fields Current number fields within cluster.
+# TYPE elasticsearch_indices_mappings_stats_fields gauge
+elasticsearch_indices_mappings_stats_fields{index="test-data-2023.01.20"} 40
+			`,
+		},
 	}
-	for ver, out := range tcs {
-		for hn, handler := range map[string]http.Handler{
-			"plain": http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-				fmt.Fprintln(w, out)
-			}),
-		} {
-			ts := httptest.NewServer(handler)
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			f, err := os.Open(tt.file)
+			if err != nil {
+				t.Fatal(err)
+			}
+			defer f.Close()
+
+			ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+				io.Copy(w, f)
+			}))
 			defer ts.Close()
 
 			u, err := url.Parse(ts.URL)
 			if err != nil {
-				t.Fatalf("Failed to parse URL: %s", err)
+				t.Fatal(err)
 			}
-			c := NewIndicesMappings(log.NewNopLogger(), http.DefaultClient, u)
-			imr, err := c.fetchAndDecodeIndicesMappings()
+
+			c := NewIndicesMappings(promslog.NewNopLogger(), http.DefaultClient, u)
 			if err != nil {
-				t.Fatalf("Failed to fetch or decode indices mappings: %s", err)
-			}
-			t.Logf("[%s/%s] All Indices Mappings Response: %+v", hn, ver, imr)
-
-			response := *imr
-			if *response["facebook"].Mappings.Properties["contact"].Properties["phone"].Type != "text" {
-				t.Errorf("Marshalling error at facebook.contact.phone")
+				t.Fatal(err)
 			}
 
-			if *response["facebook"].Mappings.Properties["contact"].Properties["email"].Fields["raw"].Type != "keyword" {
-				t.Errorf("Marshalling error at facebook.contact.email.raw")
+			if err := testutil.CollectAndCompare(c, strings.NewReader(tt.want)); err != nil {
+				t.Fatal(err)
 			}
-
-			if *response["facebook"].Mappings.Properties["name"].Type != "text" {
-				t.Errorf("Marshalling error at facebook.name")
-			}
-
-			if *response["facebook"].Mappings.Properties["name"].Fields["raw"].Type != "keyword" {
-				t.Errorf("Marshalling error at facebook.name.raw")
-			}
-
-			if *response["twitter"].Mappings.Properties["email"].Type != "keyword" {
-				t.Errorf("Marshalling error at twitter.email")
-			}
-
-			if *response["twitter"].Mappings.Properties["phone"].Type != "keyword" {
-				t.Errorf("Marshalling error at twitter.phone")
-			}
-
-		}
+		})
 	}
-}
-
-func TestIndexMappingFieldCount(t *testing.T) {
-
-	testIndexNumFields := 40.0
-	testIndexName := "test-data-2023.01.20"
-
-	rawMapping := `{
-			"test-data-2023.01.20": {
-				"mappings": {
-					"properties": {
-						"data": {
-							"type": "object",
-							"properties": {
-								"field1": {
-									"type": "text",
-									"fields": {
-										"keyword": {
-											"type": "keyword",
-											"ignore_above": 256
-										}
-									}
-								},
-								"field10": {
-									"type": "long"
-								},
-								"field2": {
-									"type": "text",
-									"fields": {
-										"keyword": {
-											"type": "keyword",
-											"ignore_above": 256
-										}
-									}
-								},
-								"field3": {
-									"type": "text",
-									"fields": {
-										"keyword": {
-											"type": "keyword",
-											"ignore_above": 256
-										}
-									}
-								},
-								"field4": {
-									"type": "text",
-									"fields": {
-										"keyword": {
-											"type": "keyword",
-											"ignore_above": 256
-										}
-									}
-								},
-								"field5": {
-									"type": "text",
-									"fields": {
-										"keyword": {
-											"type": "keyword",
-											"ignore_above": 256
-										}
-									}
-								},
-								"field6": {
-									"type": "text",
-									"fields": {
-										"keyword": {
-											"type": "keyword",
-											"ignore_above": 256
-										}
-									}
-								},
-								"field7": {
-									"type": "text",
-									"fields": {
-										"keyword": {
-											"type": "keyword",
-											"ignore_above": 256
-										}
-									}
-								},
-								"field8": {
-									"type": "text",
-									"fields": {
-										"keyword": {
-											"type": "keyword",
-											"ignore_above": 256
-										}
-									}
-								},
-								"field9": {
-									"type": "long"
-								}
-							}
-						},
-						"data2": {
-							"properties": {
-								"field1": {
-									"type": "text",
-									"fields": {
-										"keyword": {
-											"type": "keyword",
-											"ignore_above": 256
-										}
-									}
-								},
-								"field2": {
-									"type": "text",
-									"fields": {
-										"keyword": {
-											"type": "keyword",
-											"ignore_above": 256
-										}
-									}
-								},
-								"field3": {
-									"type": "text",
-									"fields": {
-										"keyword": {
-											"type": "keyword",
-											"ignore_above": 256
-										}
-									}
-								},
-								"field4": {
-									"type": "text",
-									"fields": {
-										"keyword": {
-											"type": "keyword",
-											"ignore_above": 256
-										}
-									}
-								},
-								"field5": {
-									"type": "text",
-									"fields": {
-										"keyword": {
-											"type": "keyword",
-											"ignore_above": 256
-										}
-									}
-								},
-								"nested_field6": {
-									"properties": {
-										"field1": {
-											"type": "text",
-											"fields": {
-												"keyword": {
-													"type": "keyword",
-													"ignore_above": 256
-												}
-											}
-										},
-										"field2": {
-											"type": "text",
-											"fields": {
-												"keyword": {
-													"type": "keyword",
-													"ignore_above": 256
-												}
-											}
-										},
-										"field3": {
-											"type": "text",
-											"fields": {
-												"keyword": {
-													"type": "keyword",
-													"ignore_above": 256
-												}
-											}
-										},
-										"field4": {
-											"type": "text",
-											"fields": {
-												"keyword": {
-													"type": "keyword",
-													"ignore_above": 256
-												}
-											}
-										},
-										"field5": {
-											"type": "long"
-										}
-									}
-								}
-							}
-						}
-					}
-				}
-			}
-		}`
-
-	for _, handler := range map[string]http.Handler{
-		"plain": http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			fmt.Fprintln(w, rawMapping)
-		}),
-	} {
-
-		ts := httptest.NewServer(handler)
-		defer ts.Close()
-
-		u, err := url.Parse(ts.URL)
-		if err != nil {
-			t.Fatalf("Failed to parse URL: %s", err)
-		}
-		c := NewIndicesMappings(log.NewNopLogger(), http.DefaultClient, u)
-		indicesMappingsResponse, err := c.fetchAndDecodeIndicesMappings()
-		if err != nil {
-			t.Fatalf("Failed to fetch or decode indices mappings: %s", err)
-		}
-
-		response := *indicesMappingsResponse
-		mapping := response[testIndexName]
-		totalFields := countFieldsRecursive(mapping.Mappings.Properties, 0)
-		if totalFields != testIndexNumFields {
-			t.Errorf("Number of actual fields in index doesn't match the count returned by the recursive countFieldsRecursive function")
-		}
-
-	}
-
 }
