@@ -25,7 +25,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/go-kit/log"
+	"github.com/prometheus/common/promslog"
 
 	"github.com/blang/semver/v4"
 )
@@ -44,7 +44,7 @@ const (
 
 type mockES struct{}
 
-func (mockES) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+func (mockES) ServeHTTP(w http.ResponseWriter, _ *http.Request) {
 
 	fmt.Fprintf(w, `{
   "name" : "%s",
@@ -120,7 +120,7 @@ func TestNew(t *testing.T) {
 	if err != nil {
 		t.Skipf("internal test error: %s", err)
 	}
-	r := New(log.NewNopLogger(), http.DefaultClient, u, 0)
+	r := New(promslog.NewNopLogger(), http.DefaultClient, u, 0)
 	if r.url != u {
 		t.Errorf("new Retriever mal-constructed")
 	}
@@ -132,7 +132,7 @@ func TestRetriever_RegisterConsumer(t *testing.T) {
 	if err != nil {
 		t.Fatalf("internal test error: %s", err)
 	}
-	retriever := New(log.NewNopLogger(), mockES.Client(), u, 0)
+	retriever := New(promslog.NewNopLogger(), mockES.Client(), u, 0)
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 	consumerNames := []string{"consumer-1", "consumer-2"}
@@ -171,7 +171,7 @@ func TestRetriever_fetchAndDecodeClusterInfo(t *testing.T) {
 	if err != nil {
 		t.Skipf("internal test error: %s", err)
 	}
-	retriever := New(log.NewNopLogger(), mockES.Client(), u, 0)
+	retriever := New(promslog.NewNopLogger(), mockES.Client(), u, 0)
 	ci, err := retriever.fetchAndDecodeClusterInfo()
 	if err != nil {
 		t.Fatalf("failed to retrieve cluster info: %s", err)
@@ -191,7 +191,7 @@ func TestRetriever_Run(t *testing.T) {
 	}
 
 	// setup cluster info retriever
-	retriever := New(log.NewLogfmtLogger(os.Stdout), mockES.Client(), u, 0)
+	retriever := New(promslog.New(&promslog.Config{Writer: os.Stdout}), mockES.Client(), u, 0)
 
 	// setup mock consumer
 	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
